requirements.txt
opencv-python==4.8.0
numpy==1.24.3
face-recognition==1.3.0
dlib==19.24.2
pillow==10.0.0
tkinter
sqlite3 